Disable implicit casts (#130)
This will make it easier to migrate to null safety which also disables
implicit casts except from dynamic. Also fix the implicit casts from
dynamic since this package has a lot of dynamic code that is harder to
read with implicit casts.

- Add argument types to some signatures that should have had them from
  the start.
- Add explicit casts.
- Extract a few new local variables to make some of the casts only need
  to happen once.
- Use `Iterable.whereType` in a loop instead of a type check and
  `continue` in the loop body.

+++ lib/src/utils.dart
@@ -79,7 +79,7 @@ String formatStr(String format, Map data) {
           result.write(padWithZeros(number, numberSize));
           break;
         case 'x':
-          var number = value.toRadixString(16);
+          var number = (value as int).toRadixString(16);
           result.write(padWithZeros(number, numberSize));
           break;
         default:
